test(categoryOrderHints): add unit test to handle empty categoryOrderHint
this is necessary since the api/platform adds an empty object as the categoryOrderHints even if there is non provided in the product that was created. Thus when you compare it to the original product there is a suddenly the categoryOrderHints field and the sdk would create an unnecessary action for this.